refactor(packages): reorganizar paquetes de vistas a 'view'
