+++ CarsAccessoriesApp/src/main/java/com/pro/carsaccessoriesapp/view/LoginFrame.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package com.pro.carsaccessoriesapp.ejem;
+package com.pro.carsaccessoriesapp.view;
 
 /**
  *
